Add script to build, run and collect benchmark results
This patch adds python scripts that help build, run and gather results
from the benchmarks. As well as a basic script to compare results from
two different runs.

It works with a `.json` file containing the benchmark names, a regex to
find the timings in the benchmark output and optional arguments that
must be provided to the benchmark binary. The `subset.json` contains
roughly 70 of the benchmarks for cuda, hip and sycl at the moment, more
work would be required to support the rest of the benchmarks. In
addition if there are failing benchmarks in the `.json` list, an
additional text file can be provided with a list of benchmarks to skip
when running all of them. Benchmarks in the text file can still be run
explicitely.

For example to run all the SYCL benchmarks and then all the CUDA
benchmarks and compare the two:

```
./autohecbench.py sycl -o sycl.csv
./autohecbench.py cuda -o cuda.csv
./autohecbench-compare.py sycl.csv cuda.csv
```

It can also be used to run a single benchmark:

```
./autohecbench.py mandelbrot-sycl --verbose
```

By default it will run a warmup iteration before running each benchmark,
and it is possible to run the benchmarks multiple times with `-r`:

```
./autohecbench.py mandelbrot-sycl -r 20 -o mandel.csv
```

And it also has options to pick the SM version or HIP architecture and a
few other parameters.